requirements.txt
aiohttp==3.6.2
asyncio==3.4.3
discord.py==1.4.1
DiscordUtils==1.2.8
requests==2.23.0